mocks for application dao moved into separate file (mock_daos.go -> mock_dao_application.go)
