+++ internal/testutils/mocks/mock_daos.go
@@ -3,8 +3,6 @@ package mocks
 import (
 	"errors"
 	"fmt"
-	"strings"
-
 	"github.com/RedHatInsights/sources-api-go/config"
 	"github.com/RedHatInsights/sources-api-go/internal/testutils/fixtures"
 	m "github.com/RedHatInsights/sources-api-go/model"
@@ -12,10 +10,6 @@ import (
 )
 
 var conf = config.Get()
-
-type MockApplicationDao struct {
-	Applications []m.Application
-}
 
 type MockEndpointDao struct {
 	Endpoints []m.Endpoint
@@ -183,6 +177,7 @@ func (a *MockApplicationTypeDao) ApplicationTypeCompatibleWithSource(_, _ int64)
 func (at *MockApplicationTypeDao) GetSuperKeyResultType(applicationTypeId int64, authType string) (string, error) {
 	panic("not needed")
 }
+
 func (a *MockApplicationTypeDao) ApplicationTypeCompatibleWithSourceType(_, _ int64) error {
 	if a.Compatible {
 		return nil
@@ -224,141 +219,6 @@ func (a *MockSourceTypeDao) Update(src *m.SourceType) error {
 
 func (a *MockSourceTypeDao) Delete(id *int64) error {
 	panic("not implemented") // TODO: Implement
-}
-
-func (a *MockApplicationDao) SubCollectionList(primaryCollection interface{}, limit, offset int, filters []util.Filter) ([]m.Application, int64, error) {
-	var applications []m.Application
-
-	switch object := primaryCollection.(type) {
-	case m.Source:
-		var sourceExists bool
-
-		for _, s := range fixtures.TestSourceData {
-			if s.ID == object.ID {
-				sourceExists = true
-			}
-		}
-		// if source doesn't exist, return Not Found Err
-		if !sourceExists {
-			return nil, 0, util.NewErrNotFound("source")
-		}
-
-		// else return list of related applications
-		for _, app := range a.Applications {
-			if object.ID == app.SourceID {
-				applications = append(applications, app)
-			}
-		}
-	}
-
-	return applications, int64(len(applications)), nil
-}
-
-func (a *MockApplicationDao) List(limit int, offset int, filters []util.Filter) ([]m.Application, int64, error) {
-	count := int64(len(a.Applications))
-	return a.Applications, count, nil
-}
-
-func (a *MockApplicationDao) GetById(id *int64) (*m.Application, error) {
-	for _, app := range a.Applications {
-		if app.ID == *id {
-			return &app, nil
-		}
-	}
-
-	return nil, util.NewErrNotFound("application")
-}
-
-func (a *MockApplicationDao) GetByIdWithPreload(id *int64, preloads ...string) (*m.Application, error) {
-	for _, app := range a.Applications {
-		if app.ID == *id {
-			for _, preload := range preloads {
-				if strings.Contains(strings.ToLower(preload), "source") {
-					app.Source = fixtures.TestSourceData[0]
-				}
-			}
-
-			return &app, nil
-		}
-	}
-
-	return nil, util.NewErrNotFound("application")
-}
-
-func (a *MockApplicationDao) Create(src *m.Application) error {
-	return nil
-}
-
-func (a *MockApplicationDao) Update(src *m.Application) error {
-	return nil
-}
-
-func (a *MockApplicationDao) Delete(id *int64) (*m.Application, error) {
-	for _, app := range a.Applications {
-		if app.ID == *id {
-			return &app, nil
-		}
-	}
-	return nil, util.NewErrNotFound("application")
-}
-
-func (a *MockApplicationDao) Tenant() *int64 {
-	tenant := int64(1)
-	return &tenant
-}
-
-func (a *MockApplicationDao) User() *int64 {
-	user := int64(1)
-	return &user
-}
-
-func (a *MockApplicationDao) DeleteCascade(applicationId int64) ([]m.ApplicationAuthentication, *m.Application, error) {
-	var application *m.Application
-	for _, app := range fixtures.TestApplicationData {
-		if app.ID == applicationId {
-			application = &app
-		}
-	}
-
-	if application == nil {
-		return nil, nil, util.NewErrNotFound("application")
-	}
-
-	return fixtures.TestApplicationAuthenticationData, application, nil
-}
-
-func (a *MockApplicationDao) Exists(applicationId int64) (bool, error) {
-	for _, application := range a.Applications {
-		if application.ID == applicationId {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-func (m *MockApplicationDao) BulkMessage(_ util.Resource) (map[string]interface{}, error) {
-	return nil, nil
-}
-
-func (m *MockApplicationDao) FetchAndUpdateBy(_ util.Resource, _ map[string]interface{}) (interface{}, error) {
-	return nil, nil
-}
-
-func (m *MockApplicationDao) ToEventJSON(_ util.Resource) ([]byte, error) {
-	return nil, nil
-}
-
-func (a *MockApplicationDao) Pause(_ int64) error {
-	return nil
-}
-
-func (a *MockApplicationDao) Unpause(_ int64) error {
-	return nil
-}
-
-func (src *MockApplicationDao) IsSuperkey(id int64) bool {
-	return false
 }
 
 func (a *MockEndpointDao) SubCollectionList(primaryCollection interface{}, limit, offset int, filters []util.Filter) ([]m.Endpoint, int64, error) {
